requirements.txt
discord.py>=2.5.2,<3
python-dotenv==1.1.0